+++ src/main/java/com/tencoding/ADayOfLearning/service/LectureService.java
@@ -29,24 +29,6 @@ public class LectureService {
 		return lectureRepository.findByLectureId(id);
 	}
 
-	/**
-	 * 강의 등록
-	 * 
-	 * @param LectureRegistarionRequestDto
-	 * @return 생성된 강의의 id
-	 */
-	public int insertLecture(LectureRegistarionRequestDto dto, int userId) {
-
-		Lecture lecture = new Lecture().builder().categoryId(dto.getCategoryId()).userId(userId).title(dto.getTitle())
-				.content(dto.getContent()).address(dto.getAddress()).addressDetail(dto.getAddressDetail())
-				.maximum(dto.getMaximum()).price(dto.getPrice()).phoneNumber(dto.getPhoneNumber()).state(1)
-				.latitude(dto.getLatitude()).longitude(dto.getLongitude()).duration(dto.getDuration()).build();
-
-		lectureRepository.insert(lecture);
-
-		return lecture.getLectureId();
-	}
-
 	// indexPage start
 //	public List<LectureListItemResponseDto> getMostLecture() {
 //		List<LectureListItemResponseDto> lectureList = lectureRepository.findMostLecture();
