+++ src/main/webapp/WEB-INF/view/business/lecture/registration.jsp
